pyproject.toml
[project]
name = "openppx"
version = "0.2.1"
description = "A lightweight skills-first agent built with Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.25.0",
    "greenlet>=3.1.1",
    "lark-oapi>=1.4.15",
    "loguru>=0.7.3",
    "python-socks>=2.5.3",
    "websockets>=12.0",
    "litellm>=1.75.5,<1.80.17",
    "openai>=1.100.2",
    "httpx>=0.27.0",
    "oauth-cli-kit>=0.1.1",
]

[project.optional-dependencies]
wecom = [
    "wecom-aibot-sdk-python>=0.1.5",
]
weixin = [
    "qrcode[pil]>=8.0",
    "pycryptodome>=3.20.0",
]

[project.scripts]
ppx = "openppx.app.cli:main"
openppx-gui-mcp = "openppx.gui.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openppx"]

[tool.hatch.build]
include = [
    "openppx/**/*.py",
    "openppx/bridge/**",
    "openppx/skills/**",
]
